requirements.txt
discord.py==1.7.2
pycoingecko==2.2.0